fix: refresh token index was not in sync
