+++ src/models/Auth/RefreshToken.ts
@@ -38,3 +38,4 @@ const RefreshTokenSchema = new Schema<IRefreshToken>({
 });
 
 export const RefreshToken = model("RefreshToken", RefreshTokenSchema);
+RefreshToken.syncIndexes();
